requirements.txt
selene>=2.0.0b14
pytest~=7.1.3
allure-pytest

requests~=2.28.1
selenium~=4.4.3
python-dotenv~=0.21.0

pydantic
Appium-Python-Client~=2.6.1